==> evaluation/03_bart_keep-all.ipynb <==
# %run 01_preprocess.ipynb
%run 00_common.ipynb

MODEL_PATTERN='keep_all'
TORGO_TRAIN_TYPE=TorgoTrainType.WORD_KEEP.value

loaded_datasets = {}

for speaker in SPEAKERS:
    file_pattern = f'torgo_xlsr_finetune_{speaker}_{MODEL_PATTERN}_{{split}}.json'

    loaded_datasets[speaker] = load_dataset('json', data_files={
        'train': os.path.join(f'data/{TORGO_TRAIN_TYPE}', file_pattern.format(split='train')),
        'valid': os.path.join(f'data/{TORGO_TRAIN_TYPE}',  file_pattern.format(split='val')),
        'test': os.path.join(f'data/{TORGO_TRAIN_TYPE}',  file_pattern.format(split='test')),
    })

for speaker, dataset in loaded_datasets.items():
    print(f"SPEAKER {speaker} - Train Set Size: {len(dataset['train'])}")
    print(f"SPEAKER {speaker} - Valid Set Size: {len(dataset['valid'])}")
    print(f"SPEAKER {speaker} - Test Set Size: {len(dataset['test'])}")

# Initialize the tokenizer
tokenizer = BartTokenizerFast.from_pretrained('facebook/bart-base')

def train_model_for_speaker(speaker_id):
    print(speaker_id)
    # Apply preprocess_function to train_data and val_data
    train_data=loaded_datasets[speaker_id]['train']
    val_data=loaded_datasets[speaker_id]['valid']
    train_data = train_data.map(preprocess_function, batched=True)
    val_data = val_data.map(preprocess_function, batched=True)

    # Prepare DataLoader for training and validation
    train_dataloader = torch.utils.data.DataLoader(train_data, batch_size=16, shuffle=True)
    val_dataloader = torch.utils.data.DataLoader(val_data, batch_size=16)

    output_dir = f"torgo_spell_correction_{MODEL_PATTERN}_{speaker_id}"
    print(output_dir)

    training_args = Seq2SeqTrainingArguments(
        output_dir=output_dir,
        evaluation_strategy="epoch",
        learning_rate=1e-5,
        per_device_train_batch_size=16,
        per_device_eval_batch_size=16,
        weight_decay=0.01,
        save_total_limit=3,
        num_train_epochs=40,
        predict_with_generate=True,
        push_to_hub=False,
        logging_steps=100
    )

    print("Train Dataset:", train_data)
    print("Validation Dataset:", val_data)
    print("Tokenizer:", tokenizer)
    print("Training Arguments:", training_args)

    model = BartForConditionalGeneration.from_pretrained('facebook/bart-base')
    device = torch.device('cuda' if torch.cuda.is_available() else 'cpu')
    model.to(device)

    # The Seq2SeqTrainer is created with the defined model, training arguments, datasets, tokenizer, and data_collator
    trainer = Seq2SeqTrainer(
        model=model,
        args=training_args,
        train_dataset=train_data,
        eval_dataset=val_data,
        tokenizer=tokenizer,
        data_collator=data_collator
    )
    trainer.train()

train_model_for_speaker('F01')

train_model_for_speaker('F03')


train_model_for_speaker('F04')


train_model_for_speaker('M01')

train_model_for_speaker('M02')


train_model_for_speaker('M03')


train_model_for_speaker('M04')


train_model_for_speaker('M05')

from transformers import BartForConditionalGeneration

SPEAKERS = ['F01', 'F03', 'F04', 'M01', 'M02', 'M03', 'M04', 'M05']
models = {}

for speaker in SPEAKERS:
    model_name = f"torgo_spell_correction_{MODEL_PATTERN}_{speaker}/checkpoint-500"
    models[speaker] = BartForConditionalGeneration.from_pretrained(model_name)

device = torch.device('cuda' if torch.cuda.is_available() else 'cpu')
for speaker, model in models.items():
    models[speaker] = model.to(device)
    models[speaker].eval()

def evaluate_speaker(model, tokenizer, speaker_id):
    predictions = []
    references = []
    test_dataset = loaded_datasets[speaker_id]['test']
    test_dataset = test_dataset.map(preprocess_function, batched=True)
    model.eval()

    for example in test_dataset:
        input_text = f"{example['predictions_phoneme']}"
        with torch.no_grad():
            input_ids = tokenizer.encode(input_text, return_tensors='pt').to(device)
            outputs = model.generate(input_ids=input_ids, max_length=max_length)

        predicted_sentence = tokenizer.decode(outputs[0], skip_special_tokens=True)

        references.append(example['references_phoneme'])
        predictions.append(predicted_sentence)

    # Verify that the number of predictions and references are the same
    if len(predictions) == len(references):
        print("Number of predictions and references are the same.")
    else:
        print("Mismatch in the number of predictions and references.")

    # Print the number of predictions and references
    print("Number of predictions:", len(predictions))
    print("Number of references:", len(references))
    # print the length of the dataset
    print("Number of rows in dataset:", len(test_dataset))

    # Assuming 'predictions' and 'references' are your sequences
    # Calculate Word Error Rate (WER)
    wer_value = wer(predictions, references)
    wer_percentage = wer_value * 100
    print(f"WER for {speaker_id}: {wer_percentage:.2f}%")

    # Calculate Character Error Rate (CER)
    cer_value = cer(predictions, references)
    cer_percentage = cer_value * 100
    print(f"CER for {speaker_id}: {cer_percentage:.2f}%")

for speaker_id, model in models.items():
    evaluate_speaker(model, tokenizer, speaker_id)
