==> evaluation/06_prepare_tatoeba_data-Copy1.ipynb <==
%run 10_ngram_common.ipynb

from datasets import load_dataset
#from g2p_en import G2p
from g2p import make_g2p
from num2words import num2words  # Import num2words from nltk
import re

# Load the Tatoeba dataset
tatoeba_dataset = load_dataset("tatoeba", 'en-mr')

# Initialize g2p converter
transducer = make_g2p('eng', 'eng-arpabet')

translations = tatoeba_dataset["train"]["translation"]

unique_phonemes = set()

with open("dataset_phonemes.txt", "w") as file:
    for translation in translations:
      sentence = translation["en"]
      sentence = remove_punctuation_and_special_characters(sentence)
      sentence = convert_numbers_to_words(sentence)
      phonemes_list = [transducer(word).output_string for word in re.findall(r'\S+', sentence)]
      all_phonemes = [item for sublist in phonemes_list for item in sublist.split()]

      # Update the set of unique phonemes excluding those containing hyphens
      unique_phonemes.update(phoneme for phoneme in all_phonemes if "-" not in phoneme)

      # Replace hyphens with letters that come after the hyphen
      all_phonemes_no_hyphen = [phoneme.split("-")[-1] if "-" in phoneme else phoneme for phoneme in all_phonemes]

      # Write all individual phonemes (without hyphens) to the file
      file.write(" ".join(all_phonemes_no_hyphen))
      file.write("\n")

# Print the vocabulary of phonemes
print("Phoneme Vocabulary:")
print(unique_phonemes)

# Remove duplicates while preserving the order

# Read the content from the file
with open("dataset_phonemes.txt", 'r') as file:
    content = file.read()

# Remove instances of consecutive triple commas (,,,)
content = content.replace(',,,', '')

# Remove instances of consecutive double commas (,,)
content = content.replace(',,', '')

# Remove individual commas
content = content.replace(',', '')
# Write the modified content back to the file
with open('dataset_phonemes.txt', 'w') as file:
    file.write(content)
    
from collections import OrderedDict

# Read the content from the file
with open("dataset_phonemes.txt", 'r') as file:
    content = file.readlines()

# Remove duplicates while preserving the order
unique_content = list(OrderedDict.fromkeys(content))

# Write the modified content back to the file
with open('dataset_phonemes.txt', 'w') as file:
    file.writelines(unique_content)
