==> evaluation/09_jamspell_ec_tatoeba.ipynb <==
%run 10_ngram_common.ipynb

%run 11_prepare_torgo_dataset.ipynb

import os
input_dataset = "dataset_phonemes.txt"
output_directory = "output_directory/"
output_model_base = "output_model.klm"

# Create the output directory if it doesn't exist
if not os.path.exists(output_directory):
    os.makedirs(output_directory)

# Ngram size options
order_trigram = 3
order_unigram = 1

# Train KENLM(DEPRECATED)

# !kenlm/build/bin/lmplz -o 3    --text dataset_phonemes.txt     --arpa output_directory/output_model.klm_trigram.arpa     --discount_fallback --skip_symbols|     kenlm/build/bin/build_binary     -T /dev/stdin output_directory/output_model.klm_trigram.arpa

# compress model file

# !kenlm/build/bin/build_binary\
#     output_directory/output_model.klm_trigram.arpa \
#     output_directory/output_model.klm_trigram.klm

arpa_path = "output_directory/output_model.klm_trigram.arpa"
vocabulary = get_vocabulary_from_arpa(arpa_path)
# Print or use the vocabulary as needed
print("Vocabulary:", vocabulary)
# Tokens to add
new_tokens = ["<pad>", "<sil>", "<spn>"]
# Tokens to exclude
tokens_to_exclude = ["<s>", "</s>"]
# Add tokens to the ARPA file excluding the specified tokens
add_tokens_to_arpa(arpa_path, new_tokens, tokens_to_exclude)
vocabulary = get_vocabulary_from_arpa(arpa_path)

# Print or use the vocabulary as needed
print("Vocabulary:", vocabulary)

%cd JamSpell
%cd build

# train JamSpell

# !./main/jamspell train ../test_data/alphabet_en.txt ../test_data/dataset_phonemes.txt model_torgo.bin


# evaluate JamSpell

# !python ../evaluate/evaluate.py\
# 	 -a ../test_data/alphabet_en.txt\
# 	 -jsp model_torgo.bin\
# 	 -mx 50000 ../test_data/dataset_phonemes.txt

import jamspell

corrector = jamspell.TSpellCorrector()
corrector.LoadLangModel('model_torgo.bin')

corrector.FixFragment('EY P')

print_dataframe_sizes(dataframes_word_no_keep_all_phoneme)

for speaker, df in dataframes_word_no_keep_all_phoneme.items():
    predictions_list = []
    references_list = []

    for index, row in df.iterrows():
        # Assuming corrector.FixFragment() is a function that corrects phoneme fragments
        corrected_predictions = corrector.FixFragment(row['predictions_phoneme'])
        predictions_list.append(corrected_predictions)

        references_list.append(row['references_phoneme'])

    print(f"predictions_list: {predictions_list}")
    # Calculate WER and CER
    wer_score = wer(predictions_list, references_list)
    cer_score = cer(predictions_list, references_list)
    wer_score*=100
    cer_score*=100

    print(f"SPEAKER: {speaker}")
    print(f"WER: {wer_score:.2f}%")
    print(f"CER: {cer_score:.2f}%")
    print("===================")

print_dataframe_sizes(dataframes_word_keep_all_phoneme)

for speaker, df in dataframes_word_keep_all_phoneme.items():
    predictions_list = []
    references_list = []

    for index, row in df.iterrows():
        # Assuming corrector.FixFragment() is a function that corrects phoneme fragments
        corrected_predictions = corrector.FixFragment(row['predictions_phoneme'])
        predictions_list.append(corrected_predictions)

        references_list.append(row['references_phoneme'])

    print(f"predictions_list: {predictions_list}")
    # Calculate WER and CER
    wer_score = wer(predictions_list, references_list)
    cer_score = cer(predictions_list, references_list)
    wer_score*=100
    cer_score*=100

    print(f"SPEAKER: {speaker}")
    print(f"WER: {wer_score:.2f}%")
    print(f"CER: {cer_score:.2f}%")
    print("===================")
