==> evaluation/10_europarl_kenlm_lm.ipynb <==
target_lang="en"  # change to your target lang
speaker = "M01"
model_name = "yip-i/torgo_xlsr_finetune-" + speaker
speaker_to_filter_out = ['MC01', 'MC02', 'MC03', 'MC04','FC01','FC02','FC03']


from datasets import load_dataset

username = "jindaznb"  # change to your username

dataset = load_dataset(f"{username}/{target_lang}_corpora_parliament_processed", split="train")

with open("text.txt", "w") as file:
  file.write(" ".join(dataset["text"]))

from huggingface_hub import notebook_login

!git config --global user.email "jindaz.work@outlook.com"
!git config --global user.name "j"

notebook_login()

from huggingface_hub import create_repo, Repository

try:
  repo = Repository(repo_path_local, clone_from=repo_path_remote)
  print(f"\nRepository {repo_path_remote} already exists.")
except:
  create_repo(repo_path_remote, exist_ok=True)
  repo = Repository(repo_path_local, clone_from=repo_path_remote)
  print("\nRepository created.")

europarl = load_dataset(f"{username}/{target_lang}_corpora_parliament_processed", split="train")
europarl

# train wordlevel ngram model

# !kenlm/build/bin/lmplz -o 3    --text europarl.txt     --arpa {repo_path_local}/output_model.klm_trigram.arpa     --discount_fallback --skip_symbols|     kenlm/build/bin/build_binary     -T /dev/stdin {repo_path_local}/output_model.klm_trigram.arpa

!kenlm/build/bin/lmplz -o 5 <"text.txt" > "5gram.arpa" --discount_fallback

!head -20 5gram.arpa

# with open("5gram.arpa", "r") as read_file, open("5gram_correct.arpa", "w") as write_file:
#   has_added_eos = False
#   for line in read_file:
#     if not has_added_eos and "ngram 1=" in line:
#       count=line.strip().split("=")[-1]
#       write_file.write(line.replace(f"{count}", f"{int(count)+1}"))
#     elif not has_added_eos and "<s>" in line:
#       write_file.write(line)
#       write_file.write(line.replace("<s>", "</s>"))
#       has_added_eos = True
#     else:
#       write_file.write(line)

if ngram_order > 1:
  !kenlm/build/bin/build_binary {repo_path_local}/output_model.klm_trigram.arpa {repo_path_local}/output_model.klm_trigram.bin

repo.push_to_hub(commit_message="Upload n-gram model")

from transformers import AutoProcessor
test_speaker = "M01"
model_user = "macarious"
model_repo = f"torgo_xlsr_finetune_{test_speaker}"
model_name = f"{model_user}/{model_repo}"
processor = AutoProcessor.from_pretrained(model_name)

vocab_dict = processor.tokenizer.get_vocab()
sorted_vocab_dict = {k.lower(): v for k, v in sorted(vocab_dict.items(), key=lambda item: item[1])}

from pyctcdecode import build_ctcdecoder

decoder = build_ctcdecoder(
    labels=list(sorted_vocab_dict.keys()),
    kenlm_model_path="5gram.arpa",
)

from transformers import Wav2Vec2ProcessorWithLM

processor_with_lm = Wav2Vec2ProcessorWithLM(
    feature_extractor=processor.feature_extractor,
    tokenizer=processor.tokenizer,
    decoder=decoder
)

from huggingface_hub import Repository

repo = Repository(local_dir="model_staging", clone_from=model_name)

processor_with_lm.save_pretrained("model_staging")

import pandas as pd

df = pd.read_csv('torgo.csv')
df['audio'] = df['audio'].apply(lambda x: f"/work/van-speech-nlp/data/torgo/{x}")
df.to_csv('torgo.csv', index=False)
