==> evaluation/19_torgo_train 16 different LMs for 16 different test speakers.ipynb <==
%run 10_ngram_common.ipynb

# !pip install -U datasets
# !pip install fsspec==2023.9.2
# pattern = "no_keep"
# pattern = "keep_all"
pattern = "new"

# Read the dataset
data_df = pd.read_csv('torgo_new.csv')
dataset_csv = load_dataset('csv', data_files='torgo_new.csv',cache_dir=None)

speakers = data_df['speaker_id'].unique()

print(f'Speakers: {", ".join(speakers)}')

atypical_speaker_texts = {}

for speaker_id in atypical_speakers:
    test_speaker=speaker_id
    valid_speaker = 'F03' if test_speaker != 'F03' else 'F04'
    train_speaker = [s for s in speakers if s not in [test_speaker, valid_speaker]]
    
    torgo_dataset = DatasetDict()
    torgo_dataset['train'] = dataset_csv['train'].filter(lambda x: x in train_speaker, input_columns=['speaker_id'])
    torgo_dataset['validation'] = dataset_csv['train'].filter(lambda x: x == valid_speaker, input_columns=['speaker_id'])
    torgo_dataset['test'] = dataset_csv['train'].filter(lambda x: x == test_speaker, input_columns=['speaker_id'])

    # print(torgo_dataset)
    if pattern == "no_keep":
        unique_texts = set(torgo_dataset['train'].unique(column='text')) | set(torgo_dataset['validation'].unique(column='text')) | set(torgo_dataset['test'].unique(column='text'))
        unique_texts_count = {}
        
        for text in unique_texts:
          unique_texts_count[text] = {'train_validation': 0, 'test': 0}
        
        for text in torgo_dataset['train']['text']:
          unique_texts_count[text]['train_validation'] += 1
        
        for text in torgo_dataset['validation']['text']:
          unique_texts_count[text]['train_validation'] += 1
        
        for text in torgo_dataset['test']['text']:
          unique_texts_count[text]['test'] += 1
        
        texts_to_keep_in_train_validation = []
        texts_to_keep_in_test = []
        for text in unique_texts_count:
          if unique_texts_count[text]['train_validation'] < text_count_threshold and unique_texts_count[text]['test'] > 0:
            texts_to_keep_in_test.append(text)
          else:
            texts_to_keep_in_train_validation.append(text)
        
        original_data_count = {'train': len(torgo_dataset['train']), 'validation': len(torgo_dataset['validation']), 'test': len(torgo_dataset['test'])}
        
        # Update the three dataset splits
        torgo_dataset['train'] = torgo_dataset['train'].filter(lambda x: x['text'] in texts_to_keep_in_train_validation)
        torgo_dataset['validation'] = torgo_dataset['validation'].filter(lambda x: x['text'] in texts_to_keep_in_train_validation)
        torgo_dataset['test'] = torgo_dataset['test'].filter(lambda x: x['text'] in texts_to_keep_in_test)
        
        print(f'Train:       {len(torgo_dataset["train"])}/{original_data_count["train"]} ({len(torgo_dataset["train"]) * 100 // original_data_count["train"]}%)')
        print(f'Validation:  {len(torgo_dataset["validation"])}/{original_data_count["validation"]} ({len(torgo_dataset["validation"]) * 100 // original_data_count["validation"]}%)')
        print(f'Test:        {len(torgo_dataset["test"])}/{original_data_count["test"]} ({len(torgo_dataset["test"]) * 100 // original_data_count["test"]}%)')
        
        print()
        torgo_dataset
    elif pattern == "keep_all":
        pass
    elif pattern == "new":
            # Update the three dataset splits (if ['test_data'] == 1, keep in test, if ['test_data'] == 0, keep in train and validation)
        torgo_dataset['train'] = torgo_dataset['train'].filter(
            lambda x: x['test_data'] == 0)
        torgo_dataset['validation'] = torgo_dataset['validation'].filter(
            lambda x: x['test_data'] == 0)
        torgo_dataset['test'] = torgo_dataset['test'].filter(
            lambda x: x['test_data'] == 1)
    
            # Drop the 'test_data' column
        torgo_dataset['train'] = torgo_dataset['train'].remove_columns([
                                                                       'test_data'])
        torgo_dataset['validation'] = torgo_dataset['validation'].remove_columns([
                                                                                 'test_data'])
        torgo_dataset['test'] = torgo_dataset['test'].remove_columns([
                                                                     'test_data'])

    
    texts = torgo_dataset['train']['text']
    
    cleaned_texts = []
    for text in texts:
        cleaned_text = ' '.join(re.sub(r'[^a-zA-Z0-9]', '', word.lower()) for word in text.split())
        cleaned_texts.append(cleaned_text)
        
    atypical_speaker_texts[speaker_id] = cleaned_texts


for speaker_id, texts in atypical_speaker_texts.items():
    print(f"{speaker_id}: {len(texts)}")

os.makedirs("torgo_lm_arpa_files", exist_ok=True)

for speaker_id, texts in atypical_speaker_texts.items():
    with open(f"torgo_lm_arpa_files/{speaker_id}_texts_{pattern}.txt", "w") as file:
        for text in texts:
            file.write(text + "\n")

arpa_dir = "torgo_lm_arpa_files"
os.makedirs(arpa_dir, exist_ok=True)

for file_name in os.listdir("torgo_lm_arpa_files"):
    if file_name.endswith(f"{pattern}.txt"):
        speaker_id = file_name.split("_")[0]
        arpa_file = f"{arpa_dir}/{speaker_id}_3gram_{pattern}.arpa"
        txt_file = f"torgo_lm_arpa_files/{file_name}"
        !kenlm/build/bin/lmplz -o 3 < "{txt_file}" > "{arpa_file}" -S 10% --discount_fallback

atypical_speakers = ['F01', 'F03', 'F04', 'M01', 'M02', 'M03', 'M04', 'M05']

for speaker in atypical_speakers:
    arpa_file = f"torgo_lm_arpa_files/{speaker}_3gram_{pattern}.arpa"
    output_file = f"torgo_lm_arpa_files/{speaker}_unigram_{pattern}.txt"

    unigrams = []

    with open(arpa_file, "r") as file:
        start_extraction = False
        for line in file:
            line = line.strip()
            if start_extraction:
                if not line:  
                    break
                parts = line.split("\t")
                unigram = parts[1]  
                unigrams.append(unigram)
            elif line.startswith("\\1-grams:"):
                start_extraction = True

    with open(output_file, "w") as file:
        for unigram in unigrams:
            file.write(unigram + "\n")
