==> evaluation/22_libri_ood_analysis-NPTORGO.ipynb <==
%run 10_ngram_common.ipynb

normal_speakers = ['MC01', 'MC02', 'MC03', 'MC04','FC01','FC02','FC03']
atypical_speakers = ['F01', 'F03', 'F04', 'M01', 'M02', 'M03', 'M04', 'M05']

# read libra

libra = load_dataset("librispeech_asr", 'clean', cache_dir="/work/van-speech-nlp/cache")

translations=libra['train.360']['text']

import nltk
total_tokens = 0

# Iterate over each sentence in the translations corpus
for sentence in translations:
    # Tokenize the sentence into individual words
    tokens = nltk.word_tokenize(sentence)
    # Add the number of tokens in the current sentence to the total count
    total_tokens += len(tokens)

# Print the total number of tokens in the translations corpus
print("Total number of tokens in corpus:", total_tokens)

cleaned_translations = []

for text in translations:
    cleaned_text = ' '.join(re.sub(r'[^a-zA-Z0-9]', '', word.lower()) for word in text.split())
    cleaned_translations.append(cleaned_text)

len(translations)

with open("DATASET_libri.txt", "w") as file:
    for text in cleaned_translations:
        file.write(text + "\n")

word_set_europarl = set()

for sentence in cleaned_translations:
    words = sentence.split()
    word_set_europarl.update(words)

len(word_set_europarl)

# train LM with LIBRA

# !kenlm/build/bin/lmplz -o 3 < DATASET_libri.txt > libri.apra --discount_fallback

# read torgo

# Read the dataset
data_df = pd.read_csv('torgo_new.csv')
dataset_csv = load_dataset('csv', data_files='torgo_new.csv')

speakers = data_df['speaker_id'].unique()

print(f'Speakers: {", ".join(speakers)}')

import re

# Define a function to clean the text
def clean_text(text):
    # Remove non-alphanumeric characters and convert to lowercase
    cleaned_text = ' '.join(re.sub(r'[^a-zA-Z0-9]', '', word.lower()) for word in text.split())
    return cleaned_text

# Apply the clean_text function to the 'text' column in data_df
data_df['text'] = data_df['text'].apply(clean_text)

data_df['text']

from collections import defaultdict

# Initialize a defaultdict to store the texts for each speaker
speaker_texts = defaultdict(list)

# Iterate through each test speaker
for test_speaker in atypical_speakers:
    valid_speaker = 'F03' if test_speaker != 'F03' else 'F04'
    train_speaker = [s for s in speakers if s not in [test_speaker, valid_speaker]]
    torgo_dataset = DatasetDict()
    torgo_dataset['train'] = dataset_csv['train'].filter(lambda x: x in train_speaker, input_columns=['speaker_id'])
    torgo_dataset['validation'] = dataset_csv['train'].filter(lambda x: x == valid_speaker, input_columns=['speaker_id'])
    torgo_dataset['test'] = dataset_csv['train'].filter(lambda x: x == test_speaker, input_columns=['speaker_id'])

    torgo_dataset['train'] = torgo_dataset['train'].filter(lambda x: x['test_data'] == 0)
    torgo_dataset['validation'] = torgo_dataset['validation'].filter(lambda x: x['test_data'] == 0)
    torgo_dataset['test'] = torgo_dataset['test'].filter(lambda x: x['test_data'] == 1)

    # Filter the DataFrame to get the data for the train split for the selected train speakers
    test_data = torgo_dataset['test']
    
    # Append the texts from the filtered DataFrame to the corresponding speaker in the speaker_texts_train dictionary
    for text in test_data['text']:
        # print(text)
        new_text=clean_text(text)
        # print(new_text)
        speaker_texts[test_speaker].append(new_text)

# # Print the number of texts collected for each speaker
for speaker_id, texts in speaker_texts.items():
    print(f"{speaker_id}: {len(texts)}")
    for text in texts[:10]:
        print(text)
    print()

# calculate unique words

unique_words_per_speaker = {}

for speaker_id, texts in speaker_texts.items():
    unique_words = set()
    for text in texts:
        words = text.split()
        for word in words:
            cleaned_word = re.sub(r'[^a-zA-Z0-9]', '', word.lower())
            if cleaned_word:  
                unique_words.add(cleaned_word) 
    unique_words_per_speaker[speaker_id] = unique_words

for speaker_id, unique_words in unique_words_per_speaker.items():
    print(f"{speaker_id}: {len(unique_words)} unique words")

unique_words_per_speaker

# vocabulary similarity

out_of_domain_words_per_speaker = {}

for speaker_id, unique_words in unique_words_per_speaker.items():
    overlap = len(unique_words & word_set_europarl)
    similarity = overlap / len(unique_words) if len(unique_words) > 0 else 0
    out_of_domain_similarity = (1 - similarity)*100

    out_of_domain_words = {word for word in unique_words if word not in word_set_europarl}
    out_of_domain_words_per_speaker[speaker_id] = out_of_domain_words

    print(f"{speaker_id}: {out_of_domain_similarity:.2f}% dissimilarity with word_set_europarl")
    print(f"{speaker_id}: out_of_domain_words count: {len(out_of_domain_words)}")
    print(f"Out of domain words for {speaker_id}: {out_of_domain_words}")
    print()

# Out of domain analysis

import kenlm

model=kenlm.Model("libri_lm_arpa_files/libri.apra") 
per=model.perplexity("your text sentance")
print(per)

per=model.perplexity("resumption of the session")
print(per)

# clean the text

import re

cleaned_speaker_texts = {}

for speaker_id, texts in speaker_texts.items():
    cleaned_texts = []
    for text in texts:
        cleaned_text = ' '.join(re.sub(r'[^a-zA-Z0-9]', '', word.lower()) for word in text.split())
        cleaned_texts.append(cleaned_text)
    cleaned_speaker_texts[speaker_id] = cleaned_texts

for speaker_id, texts in cleaned_speaker_texts.items():
    print(f"Speaker {speaker_id}:")
    for text in texts[:2]:  
        print(text)
    print()

# Initialize a dictionary to store perplexities per speaker
perplexities_per_speaker = {}

# Calculate perplexity for each speaker
for speaker_id, texts in cleaned_speaker_texts.items():
    perplexities = []
    for text in texts:
        perplexity = model.perplexity(text)
        perplexities.append(perplexity)
    avg_perplexity = np.mean(perplexities)
    perplexities_per_speaker[speaker_id] = avg_perplexity

# Print average perplexity for each speaker
for speaker_id, avg_perplexity in perplexities_per_speaker.items():
    print(f"Speaker {speaker_id}: Average Perplexity: {avg_perplexity}")

# output unigram

arpa_file = f"libri_lm_arpa_files/libri.apra"
output_file = f"libri_lm_arpa_files/unigram.txt"

unigrams = []

with open(arpa_file, "r") as file:
    start_extraction = False
    for line in file:
        line = line.strip()
        if start_extraction:
            if not line:  
                break
            parts = line.split("\t")
            unigram = parts[1]  
            unigrams.append(unigram)
        elif line.startswith("\\1-grams:"):
            start_extraction = True

with open(output_file, "w") as file:
    for unigram in unigrams:
        file.write(unigram + "\n")
